4_customization/pyproject.toml
[tool.poetry]
name = "4-customization"
version = "0.1.0"
description = ""
authors = ["Michal Ustanik <michal.ustanik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
openai = "^1.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
